+++ examples/requirements-and-tests/src/language-server/requirements-and-tests-language-module.ts
@@ -1,49 +1,10 @@
 import {
     createDefaultModule, createDefaultSharedModule, DefaultSharedModuleContext, inject,
-    LangiumServices, LangiumSharedServices, Module, PartialLangiumServices
+    LangiumSharedServices
 } from 'langium';
 import { RequirementsAndTestsGeneratedSharedModule, RequirementsGeneratedModule, TestsGeneratedModule } from './generated/module';
-import { RequirementsLanguageValidationRegistry, RequirementsLanguageValidator, TestsLanguageValidationRegistry, TestsLanguageValidator } from './requirements-and-tests-language-validator';
-
-/**
- * Declaration of custom services - add your own service classes here.
- */
- export type RequirementsLanguageAddedServices = {
-    validation: {
-        RequirementsLanguageValidator: RequirementsLanguageValidator
-    }
-}
-export type TestsLanguageAddedServices = {
-    validation: {
-        TestsLanguageValidator: TestsLanguageValidator
-    }
-}
-
-/**
- * Union of Langium default services and your custom services - use this as constructor parameter
- * of custom service classes.
- */
- export type RequirementsLanguageServices = LangiumServices & RequirementsLanguageAddedServices
- export type TestsLanguageServices = LangiumServices & TestsLanguageAddedServices
-
-/**
- * Dependency injection module that overrides Langium default services and contributes the
- * declared custom services. The Langium defaults can be partially specified to override only
- * selected services, while the custom services must be fully specified.
- */
- export const RequirementsLanguageModule: Module<RequirementsLanguageServices, PartialLangiumServices & RequirementsLanguageAddedServices> = {
-    validation: {
-        ValidationRegistry: (services) => new RequirementsLanguageValidationRegistry(services),
-        RequirementsLanguageValidator: () => new RequirementsLanguageValidator()
-    }
-};
-
-export const TestsLanguageModule: Module<TestsLanguageServices, PartialLangiumServices & TestsLanguageAddedServices> = {
-    validation: {
-        ValidationRegistry: (services) => new TestsLanguageValidationRegistry(services),
-        TestsLanguageValidator: () => new TestsLanguageValidator()
-    }
-};
+import { RequirementsLanguageModule, RequirementsLanguageServices } from './requirements-language-module';
+import { TestsLanguageModule, TestsLanguageServices } from './tests-language-module';
 
 /**
  * Create the full set of services required by Langium.
